add and refactor mcv for getting users
