+++ db/upload.js
@@ -1,0 +1,16 @@
+const multer = require("multer");
+const fs = require("fs");
+const path = require("path");
+
+const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+      cb(null, "./db/uploads");
+    },
+    filename: (req, file, cb) => {
+      cb(null, file.fieldname + "-" + Date.now());
+    },
+  });
+  
+const upload = multer({ storage: storage });
+
+module.exports = upload;
