+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ./db/app.js"
+    "start": "node ./listen.js"
   },
   "keywords": [],
   "author": "",
